# telock/public/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
        "/api/students/create": {
            "post": {
                "summary": "Új diák létrehozása",
                "description": "Létrehoz egy új diákot a megadott adatokkal.",
                "tags": ["Students"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "student_id": {"type": "string", "example": "OM56555"},
                                    "full_name": {"type": "string", "example": "fasz kalap"},
                                    "class": {"type": "string", "example": "12.A"},
                                    "rfid_tag": {"type": "string", "example": "58D90E3E"}
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {"description": "Sikeresen létrehozott diák"},
                    "400": {"description": "Hiányzó kötelező mezők"},
                    "500": {"description": "Hiba a diák létrehozása során"},
                    "405": {"description": "A HTTP metódus nem engedélyezett"}
                }
            }
        },
        "/api/students/read": {
            "get": {
                "summary": "Diákok lekérése",
                "description": "A diákok adatainak lekérése.",
                "tags": ["Students"],
                "responses": {
                    "200": {"description": "Sikeres lekérés"},
                    "405": {"description": "A HTTP metódus nem engedélyezett"},
                    "500": {"description": "Hiba az adatok lekérésénél"}
                }
            }
        },
        "/api/students/update": {
            "put": {
                "summary": "Diák adatainak frissítése",
                "description": "A diák adatainak frissítésére szolgáló API végpont.",
                "tags": ["Students"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "student_id": {"type": "string", "example": "OM44444"},
                                    "full_name": {"type": "string", "example": "Pál Edvin"},
                                    "class": {"type": "string", "example": "13.D"},
                                    "rfid_tag": {"type": "string", "example": "53D00E3E"}
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {"description": "Sikeres frissítés VAGY Sikeres diák és szekrény kapcsolat frissítés"},
                    "400": {"description": "Hiányzó vagy érvénytelen mezők"},
                    "404": {"description": "A diák nem található VAGY A diák és szekrény kapcsolat nem található"},
                    "500": {"description": "Hiba történt a diák adatainak frissítése során"}
                }
            }
        },
        "/api/students/delete": {
            "delete": {
                "summary": "Diák törlése",
                "description": "A diák törlésére szolgáló API végpont.",
                "tags": ["Students"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "student_id": {"type": "string", "example": "OM56555"}
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {"description": "Sikeres törlés"},
                    "400": {"description": "Hiányzó student_id"},
                    "404": {"description": "A diák nem található"},
                    "405": {"description": "A HTTP metódus nem engedélyezett"},
                    "500": {"description": "Hiba történt a diák törlése során"}
                }
            }
        }, "/api/students/setStudentGroups": {
            "post": {
                "summary": "Egy diák csoportokba rendezése",
                "description": "Új diák létrehozásánál vagy egy meglévő diák adatainak frissítésekor, ez a végpont csoportokba rendezi a diákot.",
                "tags": ["Students"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "student_id": {"type": "string", "example": "OM56555"}
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200":{"description": "Nem találtam csoport egyezést"},    
                    "201": {"description": "Sikeres csoportba rendezés"},
                    "400": {"description": "Hiányzó student_id"},
                    "404": {"description": "A diák nem található"},   
                    "500": {"description": "Hiba az adatok feltöltésekor"},
                    "405": {"description": "A metódus nem követhető"}
                }
            }
        }
       
    },
    "components": {
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "id": {"type": "integer", "example": 1},
                    "name": {"type": "string", "example": "John Doe"},
                    "email": {"type": "string", "format": "email", "example": "johndoe@example.com"}
                }
            }
        }
    }
}